+++ StorageHandler/Features/Unit/DeleteUnit/DeleteByIdUnitHandler.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
+using FluentValidation;
 using MediatR;
+using StorageHandler.Features.EntranceBucket;
 using StorageHandler.Features.Unit.Dto;
 using StorageHandler.Utils.Data;
 using StorageHandler.Utils.Exceptions;
@@ -11,12 +13,14 @@ public class DeleteByIdUnitHandler : IRequestHandler<DeleteByIdUnitCommand, Unit
     private readonly IMapper _mapper;
     private readonly ITransactionWrapper _wrapper;
     private readonly IUnitRepository _repository;
+    private readonly IEntranceBucketRepository _bucketRepository;
 
-    public DeleteByIdUnitHandler(IMapper mapper, IUnitRepository repository, ITransactionWrapper wrapper)
+    public DeleteByIdUnitHandler(IMapper mapper, IUnitRepository repository, ITransactionWrapper wrapper, IEntranceBucketRepository bucketRepository)
     {
         _mapper = mapper;
         _repository = repository;
         _wrapper = wrapper;
+        _bucketRepository = bucketRepository;
     }
 
     public async Task<UnitResponseDto> Handle(DeleteByIdUnitCommand request, CancellationToken cancellationToken)
@@ -32,6 +36,8 @@ public class DeleteByIdUnitHandler : IRequestHandler<DeleteByIdUnitCommand, Unit
 
         if (unit == null)
             throw ExceptionUtils.GetNotFoundException($"Unit with id {id} is not found");
+        if (await _bucketRepository.ExistsBucketByUnitIdAsync(id))
+            throw new ValidationException("You cannot delete this unit, because they used in buckets");
 
         var deletedUnit = DeleteByUnit(unit);
         await _repository.SaveChangesAsync(cancellationToken);
